Split Api module
Problem:
Api module became too big and it's not very convenient to work with it.

Solution:
Move type definitions from Api to Api.Type and content definitions to Api.Content.
