+++ algorand-sdk.cabal
@@ -41,7 +41,9 @@ library
       Data.Algorand.Transaction.Signed
       Network.Algorand.Node
       Network.Algorand.Node.Api
+      Network.Algorand.Node.Api.Content
       Network.Algorand.Node.Api.Json
+      Network.Algorand.Node.Api.Type
       Network.Algorand.Node.Util
   other-modules:
       Paths_algorand_sdk
